Improve password validation on frontend. Add service endpoint for fetching urgency by id so we can validate its status on activity creation. Fix ui issues with urgencies list form. Ensure we always have request id, even in recurrent job which are going in background.

+++ api/activity/internal/service/firestore_service_test.go
@@ -13,9 +13,9 @@ func TestFirestoreService_ListByUrgency(t *testing.T) {
 	t.Parallel()
 	log := utils.NewTestLogger()
 	fake := firestoretest.NewFake().WithCollection("activities", []map[string]interface{}{
-		{"id": uint(1), "urgency_id": uint(2), "employee_id": uint(5), "description": "A", "created_at": "2025-01-02T10:00:00Z"},
-		{"id": uint(2), "urgency_id": uint(3), "employee_id": uint(6), "description": "B", "created_at": "2025-01-03T10:00:00Z"},
-		{"id": uint(3), "urgency_id": uint(2), "employee_id": uint(7), "description": "C", "created_at": "2025-01-04T10:00:00Z"},
+		{"id": int64(1), "urgency_id": int64(2), "employee_id": int64(5), "description": "A", "created_at": "2025-01-02T10:00:00Z"},
+		{"id": int64(2), "urgency_id": int64(3), "employee_id": int64(6), "description": "B", "created_at": "2025-01-03T10:00:00Z"},
+		{"id": int64(3), "urgency_id": int64(2), "employee_id": int64(7), "description": "C", "created_at": "2025-01-04T10:00:00Z"},
 	})
 	svc := NewFirebaseReadService(fake, log)
 
@@ -28,9 +28,9 @@ func TestFirestoreService_ListAll(t *testing.T) {
 	t.Parallel()
 	log := utils.NewTestLogger()
 	fake := firestoretest.NewFake().WithCollection("activities", []map[string]interface{}{
-		{"id": uint(1), "urgency_id": uint(2), "employee_id": uint(5), "description": "A", "created_at": "2025-01-02T10:00:00Z"},
-		{"id": uint(2), "urgency_id": uint(3), "employee_id": uint(6), "description": "B", "created_at": "2025-01-03T10:00:00Z"},
-		{"id": uint(3), "urgency_id": uint(2), "employee_id": uint(7), "description": "C", "created_at": "2025-01-04T10:00:00Z"},
+		{"id": int64(1), "urgency_id": int64(2), "employee_id": int64(5), "description": "A", "created_at": "2025-01-02T10:00:00Z"},
+		{"id": int64(2), "urgency_id": int64(3), "employee_id": int64(6), "description": "B", "created_at": "2025-01-03T10:00:00Z"},
+		{"id": int64(3), "urgency_id": int64(2), "employee_id": int64(7), "description": "C", "created_at": "2025-01-04T10:00:00Z"},
 	})
 	svc := NewFirebaseReadService(fake, log)
